Make converters compatible.
