+++ lib/temporal/client/converter/json.rb
@@ -1,9 +1,10 @@
+require 'temporal/client/converter/base'
 require 'temporal/json'
 
 module Temporal
   module Client
     module Converter
-      class JSON
+      class JSON < Base
         ENCODING = 'json/plain'.freeze
 
         def encoding
